Discover Torch Arrows after Enchanter advancement
I decided that discovering torches, which happens upon making a stone
pick, was too easy of a pre-condition for Torch Arrows.

Since Torch Arrows require a bow with the Flame enchantment to have an
effect, it makes sense to require the player to have enchanted something
before they get the recipe for Torch Arrows.