Add License Terms
I welcome people using this plugin, but I'm not okay with people trying
to sell it or sell similar clones without making arrangements with me
first.

I may make the license more permissive in the future, but for now, usage
is free, including in servers that have paid components (assuming they
are in compliance with Mojang's EULA), but changes, redistribution,
sales, etc. without my written consent are not.