Streamline CI/CD Pipeline
The generated changelog only reports on closed issues, which means it's
only useful after a merge to master (because that's when commit comments
like `Fixes #XX` close GitHub issues).

As such, the changelog should only be created if a release is allowed.
Since this means that multiple steps are now gated by whether or not
this workflow should produce a release,
I've added an extra step that computes and saves that boolean to an
environment variable.