Add tests and stubs for Wither Arrows
This commit lays out the methods that need to be implemented for the
Wither Arrows described in issue #9.

Furthermore, the Wither Arrow already has unit tests that can be used to
verify correct implementation. These tests will automatically run
upon attempting to build the plugin with `.\gradlew build`. If any test
fails, the build will be aborted with a message indicating why the
WitherArrow was not implemented correctly.

Once all tests pass, push up the results so I can review it before
a merge to `master`.