chore: Update to support Minecraft 1.19
- Renamed the package space to use my new personal branding (jhale.dev)
- Updated dependencies to base off of Minecraft 1.19

NOTE: One of the SniperArrowTests refuses to complete, so it has been
commented out. Interestingly the build only hangs on that one test,
unless I remove the `MockBukkit.unmock()` call, in which case it hangs
in a later test. The issue seems similar to this bug report on
MockBukkit, which is marked as resolved, but the associated PRs appear
to have been closed OR contain reports that the issue persists:
https://github.com/MockBukkit/MockBukkit/issues/263

I don't have the time to resolve the hanging test issue right now, so
though I hate to commit a commented test, that's what's happened after
manual testing to verify that the Sniper Arrow still works as expected.